+++ change_log.md
@@ -1,4 +1,9 @@
 
+# 0.1.9
+
+* Bug when reading sas7bcat files [#4](https://github.com/Roche/pyreadstat/issues/4) fixed in 
+readstat [#152](https://github.com/WizardMac/ReadStat/issues/152). Readstat src files updated with latest version
+as of 12.10.2018 (latest commit: 4c798478ed1eff35ed16c067a3b0b3fd0a9a0561)
 
 # 0.1.8
 * Installation on mac fixed (static linking iconv, [#2](https://github.com/Roche/pyreadstat/issues/2))